requirements.txt
# DPG - Decentralized Payment Gateway
# Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
cryptography==46.0.3

# Blockchain & Crypto
web3==7.14.0
eth-account==0.13.7

# Payment Processing
stripe==7.4.0

# API Clients
httpx==0.25.1
aiohttp==3.9.1
requests==2.31.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
sentry-sdk==1.38.0
